requirements.txt


## user defined
pandas-profiling
certifi
ipywidgets
great_expectations
google-cloud-bigquery==2.34.1
fastapi
uvicorn
pytest
python-dotenv

